+++ tests/Tests.fs
@@ -1,7 +1,0 @@
-module DependencyConsole.Tests
-
-open Expecto
-
-[<EntryPoint>]
-let main argv =
-    Tests.runTestsInAssembly defaultConfig argv